Complete testing and change main module

+++ Поиск подстроки/search.h
@@ -1,4 +1,4 @@
 #include <stdio.h>
 #include <string.h>
 
-size_t find(char string[], char pattern[]);
+int find(char string[], char pattern[]);
